Move lexing services into services directory
